Add support for optional slaves to multi device
The "multi" ALSA device requires all defined slaves to be present.
That's typically fine for hardware devices that do not vanish when
turned off, such as sound cards. However, for Bluetooth devices, for
example, that is not the case. Those will fail to open and, hence, cause
the entire multi device to fail. That in turn stops any sound from being
emitted.
With this change we enhance said multi devices with support for optional
slaves. This optionality has to be opted in explicitly by clients, by
setting the "comment" field of the slave to "optional":

  > type multi
  > ...
  > slaves.c.pcm "bluetooth"
  > slaves.c.channels 2;
  > slaves.c.comment "optional";
  > bindings.6.slave c;
  > bindings.6.channel 0;
  > ...

If that is done, a failure to open such a slave that is caused by the
device not being found will no longer lead to overall failure of the
open operation. Instead, it will simply be ignored and the remaining
slaves be used.